discounts done

+++ app/Http/Controllers/SellerController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Discount;
 use App\Models\Item;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -125,6 +126,11 @@ class SellerController extends Controller
 
     public function deleteItemProcess($itemID) {
         $item = Item::find($itemID);
+
+        if (!$item) {
+            return redirect('seller')->with('error', "Item with id $itemID not found");
+        }
+
         $item->delete();
 
         return redirect('seller')->with('success', "Item $item->item_name with id $item->item_id deleted");
@@ -132,8 +138,95 @@ class SellerController extends Controller
 
     public function restoreItemProcess($itemID) {
         $item = Item::withTrashed()->find($itemID);
+
+        if (!$item) {
+            return redirect('seller')->with('error', "Item with id $itemID not found");
+        }
+
         $item->restore();
 
         return redirect('seller')->with('success', "Item $item->item_name with id $item->item_id restored");
+    }
+
+
+    public function masterDiscountProcess(Request $req) {
+        if ($req->has('discount_add')) {
+            // add discount
+            $req->validate([
+                'discount_name' => 'required|string|max:255',
+                'discount_item_id' => 'required|exists:items,item_id',
+                'discount_amount' => 'required|numeric|min:1|max:100',
+                'discount_start_date' => 'required|before:discount_end_date',
+                'discount_end_date' => 'required|date',
+            ], [
+                'discount_name.required' => ':attribute is required.',
+                'discount_name.string' => ':attribute must be a string.',
+                'discount_name.max' => ':attribute may not be greater than :max characters.',
+                'discount_item_id.required' => ':attribute is required.',
+                'discount_item_id.exists' => 'The selected :attribute is invalid.',
+                'discount_amount.required' => ':attribute is required.',
+                'discount_amount.numeric' => ':attribute must be a number.',
+                'discount_amount.min' => ':attribute must be at least :min.',
+                'discount_amount.max' => ':attribute must be at most :max.',
+                'discount_start_date.required' => ':attribute is required.',
+                'discount_end_date.required' => ':attribute is required.',
+            ], [
+                'discount_name' => 'Discount Name',
+                'discount_item_id' => 'Item',
+                'discount_amount' => 'Discount Amount',
+                'discount_start_date' => 'Start Date',
+                'discount_end_date' => 'End Date',
+            ]);
+
+            $discount = Discount::create([
+                "discount_name" => $req->discount_name,
+                "discount_item_id" => $req->discount_item_id,
+                "discount_amount" => $req->discount_amount,
+                "discount_start_date" => $req->discount_start_date,
+                "discount_end_date" => $req->discount_end_date,
+            ]);
+
+            return redirect('seller')->with('success', "Discount $discount->discount_name for item ".  $discount->Item->item_name . " with id $discount->discount_id added");
+        } else if ($req->has('discount_update')) {
+            // update discount
+
+            $req->validate([
+                'discount_name' => 'required|string|max:255',
+                'discount_item_id' => 'required|exists:items,item_id',
+                'discount_amount' => 'required|numeric|min:1|max:100',
+                'discount_start_date' => 'required|before:discount_end_date',
+                'discount_end_date' => 'required|date',
+            ], [
+                'discount_name.required' => ':attribute is required.',
+                'discount_name.string' => ':attribute must be a string.',
+                'discount_name.max' => ':attribute may not be greater than :max characters.',
+                'discount_item_id.required' => ':attribute is required.',
+                'discount_item_id.exists' => 'The selected :attribute is invalid.',
+                'discount_amount.required' => ':attribute is required.',
+                'discount_amount.numeric' => ':attribute must be a number.',
+                'discount_amount.min' => ':attribute must be at least :min.',
+                'discount_amount.max' => ':attribute must be at most :max.',
+                'discount_start_date.required' => ':attribute is required.',
+                'discount_end_date.required' => ':attribute is required.',
+            ], [
+                'discount_name' => 'Discount Name',
+                'discount_item_id' => 'Item',
+                'discount_amount' => 'Discount Amount',
+                'discount_start_date' => 'Start Date',
+                'discount_end_date' => 'End Date',
+            ]);
+
+            $discount = Discount::find($req->discount_id);
+
+            $discount->discount_name = $req->discount_name;
+            $discount->discount_item_id = $req->discount_item_id;
+            $discount->discount_amount = $req->discount_amount;
+            $discount->discount_start_date = $req->discount_start_date;
+            $discount->discount_end_date = $req->discount_end_date;
+
+            $discount->save();
+
+            return redirect('seller')->with('success', "Discount $discount->discount_name for item ".  $discount->Item->item_name . " with id $discount->discount_id updated");
+        }
     }
 }
